requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
Django==3.2.10
django-extensions==3.1.4
djangorestframework==3.11.2
fastapi==0.110.0
idna==3.6
pillow==10.3.0
psycopg2-binary==2.9.10
psycopg2==2.8.3
pydantic==2.6.4
pydantic_core==2.16.3
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.0
starlette==0.36.3
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
werkzeug==0.16.0
